+++ contracts/TellorGetters.sol
@@ -25,7 +25,7 @@ contract TellorGetters {
     * @param _spender address
     * @return Returns the remaining allowance of tokens granted to the _spender from the _user
     */
-    function allowance(address _user, address _spender) external view returns (uint) {
+    function allowance(address _user, address _spender) external view returns (uint256) {
         return tellor.allowance(_user, _spender);
     }
 
@@ -35,7 +35,7 @@ contract TellorGetters {
     * @param _amount uint of amount
     * @return true if the user is alloed to trade the amount specified
     */
-    function allowedToTrade(address _user, uint _amount) external view returns(bool) {
+    function allowedToTrade(address _user, uint256 _amount) external view returns (bool) {
         return tellor.allowedToTrade(_user, _amount);
     }
 
@@ -44,7 +44,7 @@ contract TellorGetters {
     * @param _user is the owner address used to look up the balance
     * @return Returns the balance associated with the passed in _user
     */
-    function balanceOf(address _user) external view returns (uint) {
+    function balanceOf(address _user) external view returns (uint256) {
         return tellor.balanceOf(_user);
     }
 
@@ -54,7 +54,7 @@ contract TellorGetters {
     * @param _blockNumber The block number when the balance is queried
     * @return The balance at _blockNumber
     */
-    function balanceOfAt(address _user, uint _blockNumber) external view returns (uint) {
+    function balanceOfAt(address _user, uint256 _blockNumber) external view returns (uint256) {
         return tellor.balanceOfAt(_user, _blockNumber);
     }
 
@@ -64,10 +64,9 @@ contract TellorGetters {
     * @param _miner address that you want to know if they solved the challenge
     * @return true if the _miner address provided solved the
     */
-    function didMine(bytes32 _challenge, address _miner) external view returns(bool) {
+    function didMine(bytes32 _challenge, address _miner) external view returns (bool) {
         return tellor.didMine(_challenge, _miner);
     }
-
 
     /**
     * @dev Checks if an address voted in a given dispute
@@ -75,10 +74,9 @@ contract TellorGetters {
     * @param _address to look up
     * @return bool of whether or not party voted
     */
-    function didVote(uint _disputeId, address _address) external view returns(bool) {
+    function didVote(uint256 _disputeId, address _address) external view returns (bool) {
         return tellor.didVote(_disputeId, _address);
     }
-
 
     /**
     * @dev allows Tellor to read data from the addressVars mapping
@@ -87,10 +85,9 @@ contract TellorGetters {
     * addressVars[keccak256("_owner")]
     * addressVars[keccak256("tellorContract")]
     */
-    function getAddressVars(bytes32 _data) external view returns(address) {
+    function getAddressVars(bytes32 _data) external view returns (address) {
         return tellor.getAddressVars(_data);
     }
-
 
     /**
     * @dev Gets all dispute variables
@@ -113,16 +110,19 @@ contract TellorGetters {
     * @return uint of fee
     * @return int count of the current tally
     */
-    function getAllDisputeVars(uint _disputeId) public view returns(bytes32, bool, bool, bool, address, address, address, uint[9] memory, int) {
+    function getAllDisputeVars(uint256 _disputeId)
+        public
+        view
+        returns (bytes32, bool, bool, bool, address, address, address, uint256[9] memory, int256)
+    {
         return tellor.getAllDisputeVars(_disputeId);
     }
-
 
     /**
     * @dev Getter function for variables for the requestId being currently mined(currentRequestId)
     * @return current challenge, curretnRequestId, level of difficulty, api/query string, and granularity(number of decimals requested), total tip for the request
     */
-    function getCurrentVariables() external view returns(bytes32, uint, uint, string memory, uint, uint) {
+    function getCurrentVariables() external view returns (bytes32, uint256, uint256, string memory, uint256, uint256) {
         return tellor.getCurrentVariables();
     }
 
@@ -131,10 +131,9 @@ contract TellorGetters {
     * @param _hash is the sha256(abi.encodePacked(_miners[2],_requestId));
     * @return uint disputeId
     */
-    function getDisputeIdByDisputeHash(bytes32 _hash) external view returns(uint) {
-        return  tellor.getDisputeIdByDisputeHash(_hash);
+    function getDisputeIdByDisputeHash(bytes32 _hash) external view returns (uint256) {
+        return tellor.getDisputeIdByDisputeHash(_hash);
     }
-
 
     /**
     * @dev Checks for uint variables in the disputeUintVars mapping based on the disuputeId
@@ -144,30 +143,27 @@ contract TellorGetters {
     * commented out under the disputeUintVars under the Dispute struct
     * @return uint value for the bytes32 data submitted
     */
-    function getDisputeUintVars(uint _disputeId, bytes32 _data) external view returns(uint) {
+    function getDisputeUintVars(uint256 _disputeId, bytes32 _data) external view returns (uint256) {
         return tellor.getDisputeUintVars(_disputeId, _data);
     }
-
 
     /**
     * @dev Gets the a value for the latest timestamp available
     * @return value for timestamp of last proof of work submited
     * @return true if the is a timestamp for the lastNewValue
     */
-    function getLastNewValue() external view returns(uint, bool) {
+    function getLastNewValue() external view returns (uint256, bool) {
         return tellor.getLastNewValue();
     }
-
 
     /**
     * @dev Gets the a value for the latest timestamp available
     * @param _requestId being requested
     * @return value for timestamp of last proof of work submited and if true if it exist or 0 and false if it doesn't
     */
-    function getLastNewValueById(uint _requestId) external view returns(uint, bool) {
+    function getLastNewValueById(uint256 _requestId) external view returns (uint256, bool) {
         return tellor.getLastNewValueById(_requestId);
     }
-
 
     /**
     * @dev Gets blocknumber for mined timestamp
@@ -175,10 +171,9 @@ contract TellorGetters {
     * @param _timestamp is the timestamp to look up blocknumber
     * @return uint of the blocknumber which the dispute was mined
     */
-    function getMinedBlockNum(uint _requestId, uint _timestamp) external view returns(uint) {
+    function getMinedBlockNum(uint256 _requestId, uint256 _timestamp) external view returns (uint256) {
         return tellor.getMinedBlockNum(_requestId, _timestamp);
     }
-
 
     /**
     * @dev Gets the 5 miners who mined the value for the specified requestId/_timestamp
@@ -186,19 +181,17 @@ contract TellorGetters {
     * @param _timestamp is the timestamp to look up miners for
     * @return the 5 miners' addresses
     */
-    function getMinersByRequestIdAndTimestamp(uint _requestId, uint _timestamp) external view returns(address[5] memory) {
+    function getMinersByRequestIdAndTimestamp(uint256 _requestId, uint256 _timestamp) external view returns (address[5] memory) {
         return tellor.getMinersByRequestIdAndTimestamp(_requestId, _timestamp);
     }
-
 
     /**
     * @dev Get the name of the token
     * return string of the token name
     */
-    function getName() external view returns(string memory) {
+    function getName() external view returns (string memory) {
         return tellor.getName();
     }
-
 
     /**
     * @dev Counts the number of values that have been submited for the request
@@ -207,27 +200,25 @@ contract TellorGetters {
     * @param _requestId the requestId to look up
     * @return uint count of the number of values received for the requestId
     */
-    function getNewValueCountbyRequestId(uint _requestId) external view returns(uint) {
+    function getNewValueCountbyRequestId(uint256 _requestId) external view returns (uint256) {
         return tellor.getNewValueCountbyRequestId(_requestId);
     }
-
 
     /**
     * @dev Getter function for the specified requestQ index
     * @param _index to look up in the requestQ array
     * @return uint of reqeuestId
     */
-    function getRequestIdByRequestQIndex(uint _index) external view returns(uint) {
+    function getRequestIdByRequestQIndex(uint256 _index) external view returns (uint256) {
         return tellor.getRequestIdByRequestQIndex(_index);
     }
-
 
     /**
     * @dev Getter function for requestId based on timestamp
     * @param _timestamp to check requestId
     * @return uint of reqeuestId
     */
-    function getRequestIdByTimestamp(uint _timestamp) external view returns(uint) {
+    function getRequestIdByTimestamp(uint256 _timestamp) external view returns (uint256) {
         return tellor.getRequestIdByTimestamp(_timestamp);
     }
 
@@ -236,19 +227,17 @@ contract TellorGetters {
     * @param _request is the hash(of string api and granularity) to check if a request already exists
     * @return uint requestId
     */
-    function getRequestIdByQueryHash(bytes32 _request) external view returns(uint) {
+    function getRequestIdByQueryHash(bytes32 _request) external view returns (uint256) {
         return tellor.getRequestIdByQueryHash(_request);
     }
-
 
     /**
     * @dev Getter function for the requestQ array
     * @return the requestQ arrray
     */
-    function getRequestQ() public view returns(uint[51] memory) {
+    function getRequestQ() public view returns (uint256[51] memory) {
         return tellor.getRequestQ();
     }
-
 
     /**
     * @dev Allowes access to the uint variables saved in the apiUintVars under the requestDetails struct
@@ -259,10 +248,9 @@ contract TellorGetters {
     * commented out under the apiUintVars under the requestDetails struct
     * @return uint value of the apiUintVars specified in _data for the requestId specified
     */
-    function getRequestUintVars(uint _requestId, bytes32 _data) external view returns(uint) {
+    function getRequestUintVars(uint256 _requestId, bytes32 _data) external view returns (uint256) {
         return tellor.getRequestUintVars(_requestId, _data);
     }
-
 
     /**
     * @dev Gets the API struct variables that are not mappings
@@ -274,10 +262,9 @@ contract TellorGetters {
     * @return uint of index in requestQ array
     * @return uint of current payout/tip for this requestId
     */
-    function getRequestVars(uint _requestId) external view returns(string memory, string memory, bytes32, uint, uint, uint) {
+    function getRequestVars(uint256 _requestId) external view returns (string memory, string memory, bytes32, uint256, uint256, uint256) {
         return tellor.getRequestVars(_requestId);
     }
-
 
     /**
     * @dev This function allows users to retireve all information about a staker
@@ -285,7 +272,7 @@ contract TellorGetters {
     * @return uint current state of staker
     * @return uint startDate of staking
     */
-    function getStakerInfo(address _staker) external view returns(uint, uint) {
+    function getStakerInfo(address _staker) external view returns (uint256, uint256) {
         return tellor.getStakerInfo(_staker);
     }
 
@@ -295,7 +282,7 @@ contract TellorGetters {
     * @param _timestamp is the timestampt to look up miners for
     * @return address[5] array of 5 addresses ofminers that mined the requestId
     */
-    function getSubmissionsByTimestamp(uint _requestId, uint _timestamp) external view returns(uint[5] memory) {
+    function getSubmissionsByTimestamp(uint256 _requestId, uint256 _timestamp) external view returns (uint256[5] memory) {
         return tellor.getSubmissionsByTimestamp(_requestId, _timestamp);
     }
 
@@ -303,7 +290,7 @@ contract TellorGetters {
     * @dev Get the symbol of the token
     * return string of the token symbol
     */
-    function getSymbol() external view returns(string memory) {
+    function getSymbol() external view returns (string memory) {
         return tellor.getSymbol();
     }
 
@@ -313,10 +300,9 @@ contract TellorGetters {
     * @param _index is the value index to look up
     * @return uint timestamp
     */
-    function getTimestampbyRequestIDandIndex(uint _requestID, uint _index) external view returns(uint) {
+    function getTimestampbyRequestIDandIndex(uint256 _requestID, uint256 _index) external view returns (uint256) {
         return tellor.getTimestampbyRequestIDandIndex(_requestID, _index);
     }
-
 
     /**
     * @dev Getter for the variables saved under the TellorStorageStruct uintVars variable
@@ -327,19 +313,17 @@ contract TellorGetters {
     * self.uintVars[keccak256("stakerCount")]
     * @return uint of specified variable
     */
-    function getUintVar(bytes32 _data) public view returns(uint) {
+    function getUintVar(bytes32 _data) public view returns (uint256) {
         return tellor.getUintVar(_data);
     }
-
 
     /**
     * @dev Getter function for next requestId on queue/request with highest payout at time the function is called
     * @return onDeck/info on request with highest payout-- RequestId, Totaltips, and API query string
     */
-    function getVariablesOnDeck() external view returns(uint, uint, string memory) {
+    function getVariablesOnDeck() external view returns (uint256, uint256, string memory) {
         return tellor.getVariablesOnDeck();
     }
-
 
     /**
     * @dev Gets the 5 miners who mined the value for the specified requestId/_timestamp
@@ -347,10 +331,9 @@ contract TellorGetters {
     * @param _timestamp is the timestamp to look up miners for
     * @return bool true if requestId/timestamp is under dispute
     */
-    function isInDispute(uint _requestId, uint _timestamp) external view returns(bool) {
+    function isInDispute(uint256 _requestId, uint256 _timestamp) external view returns (bool) {
         return tellor.isInDispute(_requestId, _timestamp);
     }
-
 
     /**
     * @dev Retreive value from oracle based on timestamp
@@ -358,18 +341,16 @@ contract TellorGetters {
     * @param _timestamp to retreive data/value from
     * @return value for timestamp submitted
     */
-    function retrieveData(uint _requestId, uint _timestamp) external view returns (uint) {
+    function retrieveData(uint256 _requestId, uint256 _timestamp) external view returns (uint256) {
         return tellor.retrieveData(_requestId, _timestamp);
     }
-
 
     /**
     * @dev Getter for the total_supply of oracle tokens
     * @return uint total supply
     */
-    function totalSupply() external view returns (uint) {
+    function totalSupply() external view returns (uint256) {
         return tellor.totalSupply();
     }
-
 
 }
